# Optimize performance