--- a/src/explorer.rs
+++ b/src/explorer.rs
@@ -46,23 +46,29 @@ pub fn explore_async(
     config: ExplorerConfig,
     parent: String,
     focused_path: Option<String>,
-    tx: Sender<Task>,
+    tx_msg_in: Sender<Task>,
+    tx_pwd_watcher: Sender<String>,
 ) {
     thread::spawn(move || {
         explore_sync(config, parent.clone(), focused_path)
-            .map(|dirbuf| {
-                tx.send(Task::new(
-                    MsgIn::Internal(InternalMsg::AddDirectory(parent, dirbuf)),
-                    None,
-                ))
-                .unwrap_or_default();
+            .map(|buf| {
+                tx_pwd_watcher
+                    .send(buf.parent().clone())
+                    .unwrap_or_default();
+                tx_msg_in
+                    .send(Task::new(
+                        MsgIn::Internal(InternalMsg::AddDirectory(parent.clone(), buf)),
+                        None,
+                    ))
+                    .unwrap_or_default();
             })
             .unwrap_or_else(|e| {
-                tx.send(Task::new(
-                    MsgIn::External(ExternalMsg::LogError(e.to_string())),
-                    None,
-                ))
-                .unwrap_or_default();
+                tx_msg_in
+                    .send(Task::new(
+                        MsgIn::External(ExternalMsg::LogError(e.to_string())),
+                        None,
+                    ))
+                    .unwrap_or_default();
             })
     });
 }
@@ -71,16 +77,24 @@ pub fn explore_recursive_async(
     config: ExplorerConfig,
     parent: String,
     focused_path: Option<String>,
-    tx: Sender<Task>,
+    tx_msg_in: Sender<Task>,
+    tx_pwd_watcher: Sender<String>,
 ) {
     let path = PathBuf::from(&parent);
-    explore_async(config.clone(), parent, focused_path, tx.clone());
+    explore_async(
+        config.clone(),
+        parent,
+        focused_path,
+        tx_msg_in.clone(),
+        tx_pwd_watcher.clone(),
+    );
     if let Some(grand_parent) = path.parent() {
         explore_recursive_async(
             config,
             grand_parent.to_string_lossy().to_string(),
             path.file_name().map(|f| f.to_string_lossy().to_string()),
-            tx,
+            tx_msg_in,
+            tx_pwd_watcher,
         );
     }
 }