--- a/src/pwd_watcher.rs
+++ b/src/pwd_watcher.rs
@@ -17,6 +17,10 @@ pub fn keep_watching(
     thread::spawn(move || loop {
         if let Ok(new_pwd) = rx_pwd_watcher.try_recv() {
             pwd = PathBuf::from(new_pwd);
+            last_modified = pwd
+                .metadata()
+                .and_then(|m| m.modified())
+                .unwrap_or(last_modified);
         } else {
             pwd.metadata()
                 .and_then(|m| m.modified())
@@ -26,13 +30,13 @@ pub fn keep_watching(
                         tx_msg_in.send(Task::new(msg, None)).unwrap_or_default();
                         last_modified = modified;
                     } else {
-                        thread::sleep(Duration::from_millis(1000));
+                        thread::sleep(Duration::from_secs(1));
                     };
                 })
                 .unwrap_or_else(|e| {
                     let msg = MsgIn::External(ExternalMsg::LogError(e.to_string()));
                     tx_msg_in.send(Task::new(msg, None)).unwrap_or_default();
-                    thread::sleep(Duration::from_millis(1000));
+                    thread::sleep(Duration::from_secs(1));
                 })
         }
     });