--- a/src/runner.rs
+++ b/src/runner.rs
@@ -75,6 +75,7 @@ pub fn run(mut app: app::App, focused_path: Option<String>) -> Result<Option<Str
         app.pwd().clone(),
         focused_path,
         tx_msg_in.clone(),
+        tx_pwd_watcher.clone(),
     );
 
     let mut hb = Handlebars::new();
@@ -156,12 +157,35 @@ pub fn run(mut app: app::App, focused_path: Option<String>) -> Result<Option<Str
                     terminal.clear()?;
                 }
 
+                app::MsgOut::ExplorePwd => {
+                    match explorer::explore_sync(
+                        app.explorer_config().clone(),
+                        app.pwd().clone(),
+                        app.focused_node().map(|n| n.relative_path().clone()),
+                    ) {
+                        Ok(buf) => {
+                            tx_pwd_watcher.send(buf.parent().clone())?;
+                            let msg = app::MsgIn::Internal(app::InternalMsg::AddDirectory(
+                                buf.parent().clone(),
+                                buf,
+                            ));
+                            tx_msg_in.send(app::Task::new(msg, None))?;
+                        }
+                        Err(e) => {
+                            let msg =
+                                app::MsgIn::External(app::ExternalMsg::LogError(e.to_string()));
+                            tx_msg_in.send(app::Task::new(msg, None))?;
+                        }
+                    }
+                }
+
                 app::MsgOut::ExplorePwdAsync => {
                     explorer::explore_async(
                         app.explorer_config().clone(),
                         app.pwd().clone(),
                         app.focused_node().map(|n| n.relative_path().clone()),
                         tx_msg_in.clone(),
+                        tx_pwd_watcher.clone(),
                     );
                 }
 
@@ -171,20 +195,12 @@ pub fn run(mut app: app::App, focused_path: Option<String>) -> Result<Option<Str
                         app.pwd().clone(),
                         app.focused_node().map(|n| n.relative_path().clone()),
                         tx_msg_in.clone(),
+                        tx_pwd_watcher.clone(),
                     );
+                    tx_pwd_watcher.send(app.pwd().clone())?;
                 }
 
                 app::MsgOut::Refresh => {
-                    if app.pwd() != last_app.pwd() {
-                        tx_pwd_watcher.send(app.pwd().clone())?;
-                        explorer::explore_recursive_async(
-                            app.explorer_config().clone(),
-                            app.pwd().clone(),
-                            app.focused_node().map(|n| n.relative_path().clone()),
-                            tx_msg_in.clone(),
-                        );
-                    };
-
                     // UI
                     terminal.draw(|f| ui::draw(f, &app, &hb))?;
                 }